Minor fixes.
Added checks for the presence of a Analytics project key and Analytics
instance. Eqatec and Everlive both renamed to the new respective values.
Removed the filter for displayed activites (was listing the activities
of the current user). The display name for a user from the Gmail
provider is null, so we show the email for such user. Added App IDs and
App secret Keys for the social login providers.